[TIMOB-19826] Added error code and moved validaton code to Analytics class
