+++ android/modules/analytics/src/java/ti/modules/titanium/analytics/AnalyticsModule.java
@@ -7,15 +7,16 @@
 package ti.modules.titanium.analytics;
 
 import java.util.HashMap;
+import java.util.Iterator;
 
 import org.appcelerator.kroll.KrollDict;
 import org.appcelerator.kroll.KrollModule;
 import org.appcelerator.kroll.annotations.Kroll;
 import org.appcelerator.titanium.TiApplication;
-import org.appcelerator.titanium.TiC;
 import org.appcelerator.titanium.TiContext;
 import org.appcelerator.titanium.util.TiConvert;
 import org.appcelerator.titanium.util.TiPlatformHelper;
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -35,6 +36,15 @@ public class AnalyticsModule extends KrollModule
 	protected static final String PROPERTY_APP_SETTINGS = "app.settings";
 	protected static final String PROPERTY_APP_USER = "app.user";
 	private APSAnalytics analytics = APSAnalytics.getInstance();
+
+	public static final int MAX_LEVELS = 5;
+	public static final int MAX_SERLENGTH = 1000;
+	public static final int MAX_KEYS = 25;
+	public static final int MAX_KEYLENGTH = 50;
+
+	public static final int SUCCESS = 0;
+	public static final int JSON_VALIDATION_FAILED = -1;
+	public static final int ANALYTICS_DISABLED = -2;
 
 	public AnalyticsModule()
 	{
@@ -92,35 +102,86 @@ public class AnalyticsModule extends KrollModule
 	    if (TiApplication.getInstance().isAnalyticsEnabled()) {
 	        if (data instanceof HashMap) {
 	            JSONObject jsonData = TiConvert.toJSON(data);
-	            if (TiConvert.validateJSON(jsonData, 0) == TiC.ERROR_CODE_NO_ERROR) {
+	            if (AnalyticsModule.validateJSON(jsonData, 0) == SUCCESS) {
 	                analytics.sendAppFeatureEvent(event, jsonData);
-	                return TiC.ERROR_CODE_NO_ERROR;
+	                return SUCCESS;
 	            } else {
 	                Log.e(TAG, "Feature event "+ event +" not conforming to recommended usage.");
-	                return TiC.ERROR_CODE_UNKNOWN;
+	                return JSON_VALIDATION_FAILED;
 	            }
 	        } else if (data != null) {
 	            try {
 	                JSONObject jsonData = new JSONObject(data.toString());
-	                if (TiConvert.validateJSON(jsonData, 0) == TiC.ERROR_CODE_NO_ERROR) {
+	                if (AnalyticsModule.validateJSON(jsonData, 0) == SUCCESS) {
 	                    analytics.sendAppFeatureEvent(event, jsonData);
-	                    return TiC.ERROR_CODE_NO_ERROR;
+	                    return SUCCESS;
 	                } else {
 	                    Log.e(TAG, "Feature event "+ event +" not conforming to recommended usage.");
-	                    return TiC.ERROR_CODE_UNKNOWN;
+	                    return JSON_VALIDATION_FAILED;
 	                }
 	            } catch (JSONException e) {
 	                Log.e(TAG, "Cannot convert data into JSON");
-	                return TiC.ERROR_CODE_UNKNOWN;
+	                return JSON_VALIDATION_FAILED;
 	            }
 	        } else {
 	            analytics.sendAppFeatureEvent(event, null);
-	            return TiC.ERROR_CODE_NO_ERROR;
+	            return SUCCESS;
 	        }
 	    } else {
 	        Log.e(TAG, "Analytics is disabled.  To enable, please update the <analytics></analytics> node in your tiapp.xml");
-	        return TiC.ERROR_CODE_UNKNOWN;
+	        return ANALYTICS_DISABLED;
 	    }
+	}
+
+	public static int validateJSON(JSONObject jsonObject, int level) {
+
+	    if (level > MAX_LEVELS) {
+	        Log.w(TAG, "Feature event cannot have more than "+ MAX_LEVELS + " nested JSONs");
+	        return JSON_VALIDATION_FAILED;
+	    }
+	    if (jsonObject == null) {
+	        return JSON_VALIDATION_FAILED;
+	    }
+	    if ((level == 0) & (jsonObject.toString().getBytes().length > MAX_SERLENGTH)) {
+	        Log.w(TAG, "Feature event cannot exceed more than "+ MAX_SERLENGTH + " total serialized bytes");
+	        return JSON_VALIDATION_FAILED;
+	    }
+	    if (jsonObject.length() > MAX_KEYS) {
+	        Log.w(TAG, "Feature event maxium keys should not exceed "+ MAX_KEYS);
+	        return JSON_VALIDATION_FAILED;
+	    }
+
+	    Iterator<String> keys = jsonObject.keys();
+
+	    while(keys.hasNext()) {
+	        String key = (String)keys.next();
+	        if (key.length() > MAX_KEYLENGTH) {
+	            Log.w(TAG, "Feature event key "+key+" length should not exceed "+MAX_KEYLENGTH+" characters");
+	            return JSON_VALIDATION_FAILED;
+	        }
+	        try {
+	            Object child;
+	            child = jsonObject.get(key);
+	            if (child instanceof JSONObject) {
+	                if (validateJSON(((JSONObject)child), level+1) != SUCCESS){
+	                    return JSON_VALIDATION_FAILED;
+	                }
+	            } else if (jsonObject.get(key) instanceof JSONArray) {
+	                JSONArray jsonArray = (JSONArray) child;
+	                for(int i=0; i< jsonArray.length(); i++) {
+	                    Object o = jsonArray.get(i);
+	                    if (o instanceof JSONObject) {
+	                        if (validateJSON(((JSONObject)o), level+1) != SUCCESS){
+	                            return JSON_VALIDATION_FAILED;
+	                        }
+	                    }
+	                }
+	            }
+	        } catch (JSONException e) {
+	            Log.w(TAG, "Unable to validate JSON: " + e);
+	        }
+	    }
+	    return SUCCESS;
 	}
 
 	@Kroll.getProperty @Kroll.method
